MAINT: Merge _compute_neighbors and _offsets_to_raveled_neighbors
These two functions essentially do the same thing. Choose the more
descriptive names and documentation from the new function but keep
the sorting and the option to specify a custom center in `selem`.

The reasoning behind supplying `image_shape` instead of `image` is that
input arguments should be as simple as possible.